refactor?(api): move placeLotManagement deserializer in the same file than it serializer
