+++ api/src/prescription/organization-place/infrastructure/serializers/jsonapi/organization-places-lot-management-serializer.js
@@ -8,4 +8,16 @@ const serialize = function (places) {
   }).serialize(places);
 };
 
-export { serialize };
+const deserialize = function (json) {
+  return {
+    organizationId: json.data.attributes['organization-id'],
+    count: json.data.attributes['count'],
+    activationDate: json.data.attributes['activation-date'],
+    expirationDate: json.data.attributes['expiration-date'],
+    reference: json.data.attributes['reference'],
+    category: json.data.attributes['category'],
+    createdBy: json.data.attributes['created-by'],
+  };
+};
+
+export { deserialize, serialize };
