+++ api/tests/prescription/organization-place/unit/infrastructure/serializers/jsonapi/organization-places-lot-managment-serializer_test.js
@@ -1,3 +1,4 @@
+import { FREE_RATE } from '../../../../../../../src/prescription/organization-place/domain/constants/organization-places-categories.js';
 import * as organizationPlacesLotCategories from '../../../../../../../src/prescription/organization-place/domain/constants/organization-places-categories.js';
 import { OrganizationPlacesLotManagement } from '../../../../../../../src/prescription/organization-place/domain/read-models/OrganizationPlacesLotManagement.js';
 import * as organizationPlaceLotManagementSerializer from '../../../../../../../src/prescription/organization-place/infrastructure/serializers/jsonapi/organization-places-lot-management-serializer.js';
@@ -66,6 +67,41 @@ describe('Unit | Serializer | JSONAPI | organization-places-lot-management-seria
 
       // then
       expect(json).to.deep.equal(expectedJSON);
+    });
+  });
+
+  describe('#deserialize', function () {
+    //given
+    const jsonOrganizationPlacesSet = {
+      data: {
+        attributes: {
+          'organization-id': 2,
+          count: 10,
+          'activation-date': '2022-01-02',
+          'expiration-date': '2023-01-01',
+          reference: 'ABC123',
+          category: FREE_RATE,
+          'created-by': '122',
+        },
+      },
+    };
+
+    const expectedJsonOrganizationPlacesSet = {
+      organizationId: 2,
+      count: 10,
+      activationDate: '2022-01-02',
+      expirationDate: '2023-01-01',
+      reference: 'ABC123',
+      category: FREE_RATE,
+      createdBy: '122',
+    };
+
+    it('should convert JSON API data into an organization place set object', function () {
+      //when
+      const organizationPlaceSet = organizationPlaceLotManagementSerializer.deserialize(jsonOrganizationPlacesSet);
+
+      //then
+      expect(organizationPlaceSet).to.be.deep.equal(expectedJsonOrganizationPlacesSet);
     });
   });
 });
